+++ Library/components/Text/index.ts
@@ -1,0 +1,1 @@
+export { Text, type TextElement, type TextPadding, type TextProps, type TextSize, type TextWeight } from "./Text";